# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Campus Swap API",
    "description": "Comprehensive API documentation for the Campus Swap backend. This guide provides all necessary schemas and endpoints for frontend integration without requiring access to the backend source code.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3001/"
    }
  ],
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "Register a new user and create a wallet.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "User registration data",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RegisterRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "Login with email and password.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Login credentials",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/LoginRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/auth/profile": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "patch": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/listings/search": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "minPrice",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "maxPrice",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "condition",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "faculty",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/listings/": {
      "post": {
        "tags": [
          "Listings"
        ],
        "description": "Create a new marketplace listing. Requires multipart/form-data for images.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "body",
            "in": "body",
            "description": "Listing details",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CreateListingRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Listing created successfully",
            "schema": {
              "$ref": "#/components/schemas/Listing"
            }
          },
          "400": {
            "description": "Bad Request - Invalid listing data"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "get": {
        "tags": [
          "Listings"
        ],
        "description": "Retrieve all listings with optional filtering.",
        "parameters": [
          {
            "name": "faculty",
            "in": "query",
            "description": "Filter by faculty",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "in": "query",
            "description": "Filter by category",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Search title and description",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sellerId",
            "in": "query",
            "description": "Filter by seller",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filter by status (e.g., ACTIVE, SOLD)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful retrieval",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Listing"
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Listing"
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/listings/{id}": {
      "get": {
        "tags": [
          "Listings"
        ],
        "description": "Retrieve full details for a specific listing.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Listing ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Listing details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Listing"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Listing"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "patch": {
        "tags": [
          "Listings"
        ],
        "description": "Update an existing listing. Only the seller can update.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Listing ID"
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "body",
            "in": "body",
            "description": "Fields to update",
            "schema": {
              "$ref": "#/components/schemas/UpdateListingRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Listing updated successfully",
            "schema": {
              "$ref": "#/components/schemas/Listing"
            }
          },
          "400": {
            "description": "Bad Request - Invalid input"
          },
          "401": {
            "description": "Unauthorized - Token missing or invalid"
          },
          "403": {
            "description": "Forbidden - Only seller can update"
          },
          "404": {
            "description": "Not Found - Listing does not exist"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/listings/{id}/purchase": {
      "post": {
        "tags": [
          "Transactions"
        ],
        "description": "Initiate a purchase for a listing.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Listing ID"
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "body",
            "in": "body",
            "description": "Purchase options",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PurchaseRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Purchase initiated successfully"
          },
          "400": {
            "description": "Bad Request - Invalid payment method or insufficient funds"
          },
          "401": {
            "description": "Unauthorized - Token missing or invalid"
          },
          "404": {
            "description": "Not Found - Listing not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/listings/{id}/status": {
      "patch": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/listings/{id}/reserve": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/listings/{id}/save": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/analytics/listings/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/api/reports/listing/{id}": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/api/bulk/listings": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "test@unn.edu.ng"
          },
          "password": {
            "type": "string",
            "example": "Password123!"
          },
          "fullName": {
            "type": "string",
            "example": "John Doe"
          },
          "phoneNumber": {
            "type": "string",
            "example": "08012345678"
          },
          "faculty": {
            "type": "string",
            "example": "Engineering"
          }
        },
        "xml": {
          "name": "RegisterRequest"
        }
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "test@unn.edu.ng"
          },
          "password": {
            "type": "string",
            "example": "Password123!"
          }
        },
        "xml": {
          "name": "LoginRequest"
        }
      },
...
      "Listing": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid-string"
          },
          "title": {
            "type": "string",
            "example": "Calculus Textbook"
          },
          "description": {
            "type": "string",
            "example": "Used but in good condition"
          },
          "price": {
            "type": "number",
            "example": 5000
          },
          "category": {
            "type": "string",
            "example": "Books"
          },
          "condition": {
            "type": "string",
            "example": "USED"
          },
          "status": {
            "type": "string",
            "example": "ACTIVE"
          },
          "faculty": {
            "type": "string",
            "example": "Physical Sciences"
          },
          "sellerId": {
            "type": "string",
            "example": "uuid-string"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "url": {
                  "type": "string",
                  "example": "/uploads/listings/image.jpg"
                }
              }
            }
          }
        },
        "xml": {
          "name": "Listing"
        }
      },
      "CreateListingRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Calculus Textbook"
          },
          "description": {
            "type": "string",
            "example": "Used but in good condition"
          },
          "price": {
            "type": "number",
            "example": 5000
          },
          "category": {
            "type": "string",
            "example": "Books"
          },
          "condition": {
            "type": "string",
            "example": "USED"
          },
          "faculty": {
            "type": "string",
            "example": "Physical Sciences"
          }
        },
        "xml": {
          "name": "CreateListingRequest"
        }
      },
...
      "UpdateListingRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 3,
            "maxLength": 100,
            "example": "Updated Textbook"
          },
          "description": {
            "type": "string",
            "minLength": 10,
            "maxLength": 1000,
            "example": "Updated description"
          },
          "price": {
            "type": "number",
            "minimum": 0,
            "example": 6000
          },
          "category": {
            "type": "string",
            "example": "Books"
          },
          "condition": {
            "type": "string",
            "enum": [
              "NEW",
              "USED",
              "FAIR"
            ],
            "example": "USED"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "ACTIVE",
              "RESERVED",
              "SOLD",
              "ARCHIVED"
            ],
            "example": "ACTIVE"
          }
        },
        "xml": {
          "name": "UpdateListingRequest"
        }
      },
      "PurchaseRequest": {
        "type": "object",
        "required": [
          "paymentMethod"
        ],
        "properties": {
          "paymentMethod": {
            "type": "string",
            "enum": [
              "WALLET",
              "PAYSTACK_DIRECT"
            ],
            "description": "Payment method for purchase",
            "example": "WALLET"
          },
          "useEscrow": {
            "type": "boolean",
            "default": true,
            "description": "Whether to use escrow for secure transaction",
            "example": true
          },
          "meetupLocation": {
            "type": "string",
            "example": "Main Campus Gate"
          },
          "meetupTime": {
            "type": "string",
            "format": "date-time",
            "example": "2026-01-15T14:30:00Z"
          }
        },
        "xml": {
          "name": "PurchaseRequest"
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header",
        "description": "Enter your bearer token in the format **Bearer <token>**"
      }
    }
  }
}